# Fix memory leak in Buffer.toString('base64url')